add user authentication
Add user and require an user session for all views except login. Item
model updated to be tied to an user, with the admin (id 1) picked for
migration.

Set up separate formatting function for decimals. Despite rounding to 3
decimal places some conversions can still be off, eg quantities in feet
showing as 1.999 ft. The quantity is rounded to 2 or 3 decimal places,
with the former picked if it is off by 0.001 or less.